+++ src/modules/boards.module.ts
@@ -1,11 +1,15 @@
 import { Module } from '@nestjs/common';
 import { CustomTypeOrmModule } from '../configs/typeorm/custom-typeorm.module';
 import { BoardController } from '../controllers/boards.controller';
-import { BoardRespository } from '../repositories/board.repository';
+import { BoardRespository } from '../models/repositories/board.repository';
 import { BoardService } from '../services/boards.service';
 
 @Module({
-  imports: [CustomTypeOrmModule.forCustomRepository([BoardRespository])],
+  imports: [
+    CustomTypeOrmModule.forCustomRepository([
+      BoardRespository
+    ]),
+  ],
   controllers: [BoardController],
   providers: [BoardService],
 })
